reorganzise C++ classes
